requirements.txt
alabaster==0.7.12
alembic==1.8.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.1.0
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
backports.zoneinfo==0.2.1
Beaker==1.11.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
bleach==5.0.1
blinker==1.5
build==0.9.0
certifi==2021.10.8
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.12
-e git+https://github.com/ckan/ckan.git@4a438423a4f5752509e812ddb1ec349adf456781#egg=ckan
ckanext-datapackage-creator==0.0.2
click==8.1.3
click-default-group==1.2.2
colorama==0.4.6
commonmark==0.9.1
cookiecutter==2.1.1
coverage==6.5.0
coveralls==3.3.1
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
docopt==0.6.2
docutils==0.17
dominate==2.7.0
entrypoints==0.4
executing==1.2.0
factory-boy==3.2.1
Faker==14.2.0
fanstatic==1.1
fastjsonschema==2.16.2
feedgen==0.9.0
Flask==2.0.3
Flask-Babel==1.0.0
Flask-DebugToolbar==0.13.1
Flask-Login==0.6.1
flask-multistatic==1.0
Flask-WTF==1.0.1
freezegun==1.2.2
frictionless==4.40.8
funcsigs==1.0.2
greenlet==1.1.2
idna==3.3
imagesize==1.4.1
importlib-metadata==4.11.3
importlib-resources==5.10.0
incremental==22.10.0
inflection==0.5.1
iniconfig==1.1.1
ipdb==0.13.9
ipykernel==6.17.0
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
isodate==0.6.1
itsdangerous==2.1.1
jedi==0.18.1
Jinja2==3.1.2
jinja2-time==0.2.0
jsonschema==4.17.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.21.0
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
lxml==4.9.1
Mako==1.2.2
Markdown==3.4.1
marko==1.2.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
mypy==0.971
mypy-extensions==0.4.3
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.3
nbformat==5.7.0
nest-asyncio==1.5.6
nose==1.3.7
notebook==6.5.2
notebook_shim==0.2.2
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pep517==0.13.0
petl==1.7.11
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip-tools==6.8.0
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
pluggy==1.0.0
polib==1.1.1
prometheus-client==0.15.0
prompt-toolkit==3.0.31
psutil==5.9.3
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycparser==2.21
Pygments==2.13.0
PyJWT==2.4.0
pyparsing==3.0.7
pyrsistent==0.19.2
pysolr==3.9.0
pytest==7.1.3
pytest-cov==3.0.0
pytest-factoryboy==2.4.0
pytest-freezegun==0.4.2
pytest-rerunfailures==10.2
pytest-split==0.8.0
python-dateutil==2.8.2
python-editor==1.0.4
python-magic==0.4.27
python-slugify==6.1.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyUtilib==6.0.0
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.2.1
redis==4.1.4
repoze.lru==0.7
repoze.who==2.3
requests==2.28.1
responses==0.21.0
rfc3986==2.0.0
rich==12.6.0
Routes==1.13
rq==1.11.0
Send2Trash==1.8.0
shellingham==1.5.0
shutilwhich==1.1.0
simpleeval==0.9.12
simplejson==3.17.6
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==5.1.1
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.41
sqlalchemy-stubs==0.4
sqlalchemy2-stubs==0.0.2a27
sqlparse==0.4.2
stack-data==0.6.0
stringcase==1.2.0
tabulate==0.9.0
terminado==0.17.0
text-unidecode==1.3
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
towncrier==22.8.0
traitlets==5.5.0
typer==0.6.1
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
unicodecsv==0.14.1
urllib3==1.26.9
validators==0.20.0
watchdog==2.1.6
wcwidth==0.2.5
webassets==2.0
webencodings==0.5.1
WebOb==1.8.7
websocket-client==1.4.1
Werkzeug==2.0.3
widgetsnbextension==4.0.3
wrapt==1.14.0
WTForms==3.0.1
zipp==3.7.0
zope.interface==5.4.0
